@@ -1,9 +1,8 @@
 package com.rahbarbazaar.homadit.android.controllers.adapters;
 
-import android.app.AlertDialog;
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,7 +11,6 @@ import com.rahbarbazaar.homadit.android.R;
 import com.rahbarbazaar.homadit.android.controllers.interfaces.GroupGoodsItemInteraction;
 import com.rahbarbazaar.homadit.android.controllers.viewholders.GroupGoodsViewHolder;
 import com.rahbarbazaar.homadit.android.models.group_goods.GroupGoodsModel;
-import com.rahbarbazaar.homadit.android.models.searchable.SearchModel;
 
 import java.util.List;
 
