@@ -1,7 +1,7 @@
 package com.rahbarbazaar.homadit.android.controllers.viewholders
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.widget.TextView
 import com.rahbarbazaar.homadit.android.R
