@@ -2,7 +2,7 @@ package com.rahbarbazaar.homadit.android.controllers.viewholders
 
 import android.annotation.SuppressLint
 import android.content.Context
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
