@@ -1,8 +1,8 @@
 package com.rahbarbazaar.homadit.android.controllers.viewholders;
 
 import android.app.AlertDialog;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.TextView;
 
