@@ -7,8 +7,8 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.ConnectivityManager;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
@@ -18,7 +18,6 @@ import android.widget.Toast;
 import com.rahbarbazaar.homadit.android.R;
 import com.rahbarbazaar.homadit.android.controllers.adapters.LottaryPastAdapter;
 import com.rahbarbazaar.homadit.android.controllers.interfaces.LottaryPastItemInteraction;
-import com.rahbarbazaar.homadit.android.models.Lottary.ActiveLinkDetail;
 import com.rahbarbazaar.homadit.android.models.Lottary.LottaryModel;
 import com.rahbarbazaar.homadit.android.models.Lottary.OldMeDetail;
 import com.rahbarbazaar.homadit.android.models.Lottary.cancel.LottaryCancelModel;
@@ -26,7 +25,6 @@ import com.rahbarbazaar.homadit.android.models.Lottary.create.LottaryCreateModel
 import com.rahbarbazaar.homadit.android.models.Lottary.old_detail.OldDetail;
 import com.rahbarbazaar.homadit.android.models.api_error.APIError422;
 import com.rahbarbazaar.homadit.android.models.api_error.ErrorUtils;
-import com.rahbarbazaar.homadit.android.models.register.GetShopId;
 import com.rahbarbazaar.homadit.android.network.Service;
 import com.rahbarbazaar.homadit.android.network.ServiceProvider;
 import com.rahbarbazaar.homadit.android.utilities.Cache;
