@@ -8,19 +8,17 @@ import android.content.IntentFilter;
 import android.net.ConnectivityManager;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.rahbarbazaar.homadit.android.R;
 import com.rahbarbazaar.homadit.android.controllers.adapters.LottaryLinkAdapter;
 import com.rahbarbazaar.homadit.android.controllers.adapters.LottaryPrizeAdapter;
 import com.rahbarbazaar.homadit.android.controllers.interfaces.LottaryLinkItemInteraction;
 import com.rahbarbazaar.homadit.android.models.Lottary.ActiveLinkDetail;
-import com.rahbarbazaar.homadit.android.models.Lottary.ActivePrize;
 import com.rahbarbazaar.homadit.android.models.Lottary.ActivePrizeDetail;
 import com.rahbarbazaar.homadit.android.models.Lottary.LottaryModel;
 import com.rahbarbazaar.homadit.android.utilities.CustomBaseActivity;
