@@ -2,12 +2,12 @@ package com.rahbarbazaar.homadit.android.ui.fragments;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -180,7 +180,7 @@ public class HistoryFragment extends Fragment implements HistoryItemInteraction 
 
 
 
-        linearLayoutManager = new LinearLayoutManager(getContext(), LinearLayout.VERTICAL,false);
+        linearLayoutManager = new LinearLayoutManager(getContext(), LinearLayoutManager.VERTICAL,false);
         rv_history.setLayoutManager(linearLayoutManager);
 
         adapter = new HistoryAdapter(history,getContext());
