@@ -2,7 +2,7 @@ package com.rahbarbazaar.homadit.android.utilities;
 
 import android.app.Application;
 import android.content.Context;
-import android.support.multidex.MultiDex;
+import androidx.multidex.MultiDex;
 
 //import com.crashlytics.android.Crashlytics;
 //import com.squareup.leakcanary.LeakCanary;
